模型原生支持function_call/requirements.txt
python-dotenv
requests
chinesecalendar
